Fold CacheDirectory within SearchableSnapshotDirectory
