--- x-pack/plugin/searchable-snapshots/src/main/java/org/elasticsearch/index/store/BaseSearchableSnapshotDirectory.java
@@ -1,109 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License;
- * you may not use this file except in compliance with the Elastic License.
- */
-package org.elasticsearch.index.store;
-
-import org.apache.lucene.store.BaseDirectory;
-import org.apache.lucene.store.IOContext;
-import org.apache.lucene.store.IndexOutput;
-import org.apache.lucene.store.SingleInstanceLockFactory;
-import org.elasticsearch.common.blobstore.BlobContainer;
-import org.elasticsearch.index.snapshots.blobstore.BlobStoreIndexShardSnapshot;
-import org.elasticsearch.index.snapshots.blobstore.BlobStoreIndexShardSnapshot.FileInfo;
-
-import java.io.FileNotFoundException;
-import java.io.IOException;
-import java.util.Collection;
-import java.util.Objects;
-import java.util.Set;
-import java.util.concurrent.atomic.AtomicBoolean;
-
-public abstract class BaseSearchableSnapshotDirectory extends BaseDirectory {
-
-    protected final BlobStoreIndexShardSnapshot snapshot;
-    protected final BlobContainer blobContainer;
-    private final AtomicBoolean closed;
-
-    public BaseSearchableSnapshotDirectory(BlobContainer blobContainer, BlobStoreIndexShardSnapshot snapshot) {
-        super(new SingleInstanceLockFactory());
-        this.snapshot = Objects.requireNonNull(snapshot);
-        this.blobContainer = Objects.requireNonNull(blobContainer);
-        this.closed = new AtomicBoolean(false);
-    }
-
-    public BlobContainer blobContainer() {
-        return blobContainer;
-    }
-
-    protected final FileInfo fileInfo(final String name) throws FileNotFoundException {
-        return snapshot.indexFiles()
-            .stream()
-            .filter(fileInfo -> fileInfo.physicalName().equals(name))
-            .findFirst()
-            .orElseThrow(() -> new FileNotFoundException(name));
-    }
-
-    @Override
-    public final String[] listAll() {
-        ensureOpen();
-        return snapshot.indexFiles().stream().map(FileInfo::physicalName).sorted(String::compareTo).toArray(String[]::new);
-    }
-
-    @Override
-    public final long fileLength(final String name) throws IOException {
-        ensureOpen();
-        return fileInfo(name).length();
-    }
-
-    @Override
-    public Set<String> getPendingDeletions() {
-        throw unsupportedException();
-    }
-
-    @Override
-    public void sync(Collection<String> names) {
-        throw unsupportedException();
-    }
-
-    @Override
-    public void syncMetaData() {
-        throw unsupportedException();
-    }
-
-    @Override
-    public void deleteFile(String name) {
-        throw unsupportedException();
-    }
-
-    @Override
-    public IndexOutput createOutput(String name, IOContext context) {
-        throw unsupportedException();
-    }
-
-    @Override
-    public IndexOutput createTempOutput(String prefix, String suffix, IOContext context) {
-        throw unsupportedException();
-    }
-
-    @Override
-    public void rename(String source, String dest) {
-        throw unsupportedException();
-    }
-
-    private static UnsupportedOperationException unsupportedException() {
-        assert false : "this operation is not supported and should have not be called";
-        return new UnsupportedOperationException("Searchable snapshot directory does not support this operation");
-    }
-
-    @Override
-    public final void close() {
-        if (closed.compareAndSet(false, true)) {
-            isOpen = false;
-            innerClose();
-        }
-    }
-
-    protected void innerClose() {}
-}
